Move tests, remove spaces
