+++ lib/tests/streamlit/runtime/credentials_test.py
@@ -16,9 +16,12 @@
 import os
 import textwrap
 import unittest
+from pathlib import Path
 from unittest.mock import MagicMock, call, mock_open, patch
 
 import pytest
+import requests_mock
+from testfixtures import tempdir
 
 from streamlit import file_util
 from streamlit.runtime.credentials import (
@@ -30,7 +33,9 @@ from streamlit.runtime.credentials import (
 
 PROMPT = "streamlit.runtime.credentials.click.prompt"
 
-mock_get_path = MagicMock(return_value="/mock/home/folder/.streamlit/credentials.toml")
+MOCK_PATH = "/mock/home/folder/.streamlit/credentials.toml"
+
+mock_get_path = MagicMock(return_value=MOCK_PATH)
 
 
 class CredentialsClassTest(unittest.TestCase):
@@ -52,7 +57,7 @@ class CredentialsClassTest(unittest.TestCase):
         """Test Credentials constructor."""
         c = Credentials()
 
-        self.assertEqual(c._conf_file, "/mock/home/folder/.streamlit/credentials.toml")
+        self.assertEqual(c._conf_file, MOCK_PATH)
         self.assertEqual(c.activation, None)
 
     @patch(
@@ -157,7 +162,7 @@ class CredentialsClassTest(unittest.TestCase):
             self.assertEqual(
                 str(e.value).split(":")[0],
                 "\nUnable to load credentials from "
-                "/mock/home/folder/.streamlit/credentials.toml.\n"
+                f"{MOCK_PATH}.\n"
                 'Run "streamlit reset" and try again.\n',
             )
 
@@ -278,7 +283,7 @@ class CredentialsClassTest(unittest.TestCase):
 
         with patch("streamlit.runtime.credentials.os.remove") as p:
             Credentials.reset()
-            p.assert_called_once_with("/mock/home/folder/.streamlit/credentials.toml")
+            p.assert_called_once_with(MOCK_PATH)
 
         self.assertEqual(c, Credentials.get_current())
 
@@ -294,6 +299,40 @@ class CredentialsClassTest(unittest.TestCase):
             p.error.assert_called_once_with(
                 "Error removing credentials file: some error"
             )
+
+    @tempdir()
+    def test_email_send(self, temp_dir):
+        """Test that saving a new Credential sends an email"""
+
+        with requests_mock.mock() as m:
+            m.post("https://api.segment.io/v1/i", status_code=200)
+            creds: Credentials = Credentials.get_current()  # type: ignore
+            creds._conf_file = str(Path(temp_dir.path) / "config.toml")
+            creds.activation = _verify_email("email@test.com")
+            creds.save()
+            last_request = m.request_history[-1]
+            assert last_request.method == "POST"
+            assert last_request.url == "https://api.segment.io/v1/i"
+            assert "'userId': 'email@test.com'" in last_request.text
+
+    @tempdir()
+    def test_email_send_exception_handling(self, temp_dir):
+        """
+        Test that saving a new Credential catches and logs failures from the segment
+        endpoint
+        """
+
+        with requests_mock.mock() as m:
+            m.post("https://api.segment.io/v1/i", status_code=403)
+            creds: Credentials = Credentials.get_current()  # type: ignore
+            creds._conf_file = str(Path(temp_dir.path) / "config.toml")
+            creds.activation = _verify_email("email@test.com")
+            with self.assertLogs(
+                "streamlit.runtime.credentials", level="ERROR"
+            ) as mock_logger:
+                creds.save()
+                assert len(mock_logger.output) == 1
+                assert "Error saving email: 403" in mock_logger.output[0]
 
 
 class CredentialsModulesTest(unittest.TestCase):
